requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
mysql-connector-python>=8.3.0
